feat: implement usePalantirWeeklyPopupState hook for managing Palantir modal state and fetching weekly stats
