+++ src/renderStatsPage.ts
@@ -4,7 +4,6 @@ export function renderStatsPage(): string {
   return renderLayout({
     title: 'Walk to Mordor – Stats',
     description: 'Weekly insights and walking statistics for your journey to Mordor',
-    stylesheets: ['/css/palantir.css'],
     headerContent: '<h1>Stats</h1>',
     mainContent: '<div data-island="StatsIsland"></div>',
   });
